backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
sqlalchemy==2.0.34

# LLM + RAG
openai==1.47.0
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
langchain-pinecone==0.1.1
langchain-text-splitters==0.3.0

# Pinecone Python SDK
pinecone==5.3.1

# Crawler e parsing
playwright==1.47.0
beautifulsoup4==4.12.3
tldextract==5.1.2
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1
PyPDF2==3.0.1

# Redis for session storage
redis==5.0.1

# Rate limiting
slowapi==0.1.9

# Background tasks
apscheduler==3.10.4
